requirements_adk.txt
# Google ADK Migration Requirements
# Core ADK dependency
google-adk>=0.3.0

# Google AI and Vertex AI
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0

# Existing dependencies (keeping for backward compatibility during migration)
pydantic>=2.0.0
pyyaml>=6.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
streamlit>=1.28.0
plotly
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0
python-multipart>=0.0.6

# Optional: Keep LangChain for gradual migration
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.2.0
langmem>=0.0.28